koordlet: add batch resource utils
