+++ pkg/util/pod_resources_utils.go
@@ -19,16 +19,9 @@ package util
 import (
 	corev1 "k8s.io/api/core/v1"
 	quotav1 "k8s.io/apiserver/pkg/quota/v1"
-
-	apiext "github.com/koordinator-sh/koordinator/apis/extension"
 )
 
 // NOTE: functions in this file can be overwritten for extension
-
-var ExtendedResourceNames = []corev1.ResourceName{
-	apiext.BatchCPU,
-	apiext.BatchMemory,
-}
 
 func GetPodMilliCPULimit(pod *corev1.Pod) int64 {
 	podCPUMilliLimit := int64(0)
